requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
whitenoise
django-filter 